pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aioghost"
version = "0.3.1"
description = "Async Python client for the Ghost Admin API"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "vitrixbot", email = "vitrixclawd@icloud.com" }
]
keywords = ["ghost", "cms", "api", "async", "asyncio", "aiohttp"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Framework :: AsyncIO",
    "Topic :: Internet :: WWW/HTTP",
    "Typing :: Typed",
]
dependencies = [
    "aiohttp>=3.8.0",
    "PyJWT>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-aiohttp>=1.0.0",
    "pytest-cov>=7.0.0",
    "aioresponses>=0.7.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/TryGhost/aioghost"
Documentation = "https://github.com/TryGhost/aioghost#readme"
Repository = "https://github.com/TryGhost/aioghost"
Issues = "https://github.com/TryGhost/aioghost/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/aioghost"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "C4", "W"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
